Creation of the Service Controllers and DAO of the Service, Creation of the User tests and error checking tests, Fixing some problems with the User Controllers and DAO and the tests, Changing the structure of the project.

+++ Controllers/User/DeleteUser.js
@@ -1,12 +1,20 @@
 import UserDAO from "../../DAO/UserDAO.js";
+import isIdValid from "../../Utilities/isIdValid.js";
 const deleteUser = async (req, res) => {
-    //TODO:Create a function to check the id is valid or not
+
     const userDAO = new UserDAO();
     const id = req.params.id;
+    //Check if the id is valid
+    if(!isIdValid(id)){res.status(404).json({error: "Invalid ID"});return;}
+
     const user = await userDAO.deleteById(id);
 
-    //TODO:check if the user is null or not
-    res.status(200).send('User deleted');
+    //Check if the user exists or not
+    if(!user){
+        res.status(404).json({error: "User not found"});
+    }
+
+    res.status(200).json({message: "User deleted successfully"});
 };
 
 export default deleteUser;
